Working on input handling
